Run 'save-logs' after test
The 'save-logs' recipe is executed as part of the test anyway - before
removing the containers. But if the test fails in CI it is very helpful
to store the logs for any leftover containers after the end of the test.

Also make sure container logs are always uploaded as artifacts.