Load operational data from /yang-modules/operational
Loading operational data in sysrepo is not as simple as 'sysrepocfg ...'
like running configuration. An application must either push the data or
provide a callback to pull the data on demand. In this case we opt for
pushing the data read from XML files found in /yang-modules/operational.
The application must then keep the sysrepo session open for the lifetime
of the data.

We include a watch.sh script that monitors the /yang-modules/operational
directory and restarts the data provider application automatically on
any changes.